API_CORE/requirements.txt
alabaster==0.7.12
asgiref==3.5.2
Babel==2.10.3
bootstrap-py==1.0.2
certifi==2022.6.15
charset-normalizer==2.0.12
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Django==4.0.5
django-cors-headers==3.13.0
django-historicalrecords==1.1
django-simple-history==3.1.1
djangorestframework==3.13.1
docutils==0.18.1
drf-yasg==1.20.0
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
imagesize==1.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pguard==0.4.0
Pygments==2.12.0
pyparsing==3.0.9
pytz==2022.1
requests==2.28.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
smmap==5.0.0
snowballstemmer==2.2.0
Sphinx==5.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
